+++ docs/source/version_history.rst
@@ -11,6 +11,8 @@ New Features
 * Create a nice copy of the header information from any (re)bdsim file when
   loading with pybdsim using only Python types.
 * New ratio plot for 2x 1D histograms. See `pybdsim.Plot.Histogram1DRatio`.
+* pybdsim.Data.TH1,2,3 now have :code:`xrange`, :code:`yrange`, and :code:`zrange` members
+  where approriate with a conveninent tuple of the range in each dimension.
 
 Bug Fixes
 ---------